+++ Sources/KeyboardKit/_Deprecated/Keyboard+Deprecated.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, renamed: "KeyboardEnabledContext")
+public typealias KeyboardEnabledState = KeyboardEnabledContext